+++ packages/blocks/core/js/wordpress/heading/index.js
@@ -1,4 +1,8 @@
 // @flow
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
@@ -15,7 +19,13 @@ export const Heading: BlockType = {
 	name: 'blockeraHeading',
 	targetBlock: 'core/heading',
 	blockeraInnerBlocks: {
-		'elements/link': sharedInnerBlocks['elements/link'],
+		'elements/link': {
+			...sharedInnerBlocks['elements/link'],
+			settings: {
+				...sharedInnerBlocks['elements/link'].settings,
+				force: false,
+			},
+		},
 	},
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;